+++ lib/amnesie/options.rb
@@ -2,6 +2,7 @@ require 'optparse'
 
 module Amnesie
   class Options
+    attr_reader :mac, :netcard
 
     def initialize(argv)
       parse(argv)
@@ -9,6 +10,14 @@ module Amnesie
 
     def parse(argv)
       OptionParser.new do |opts|
+
+        opts.on("-m", "--mac", "What is my MAC address?") do
+          @mac = true
+        end
+
+        opts.on("-n", "--net-card NAME", "The name of the card to use") do |net|
+          @netcard = net
+        end
 
         opts.on("-h", "--help", "Show this message") do
           puts opts