+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.0.1, release 2020-05-13
+* Can randomize a mac address
+* Restart services (tor|dhcpcd)
+* Kill dhcpcd
 * Add -h|--help
 * Base project